+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -64,6 +64,7 @@ Provider = Literal[
     "reson8",
     "meta",
     "greenpt",
+    "typesafe",
     "slack",
     "orq",
 ]