+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -212,6 +212,7 @@ DeploymentsModelType = Literal[
     "rerank",
     "ocr",
     "moderation",
+    "classify",
     "vision",
 ]
 r"""The modality of the model"""
@@ -729,6 +730,7 @@ DeploymentsProvider = Literal[
     "reson8",
     "meta",
     "greenpt",
+    "typesafe",
     "slack",
     "orq",
 ]