+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentstreamop.py
@@ -2163,6 +2163,7 @@ DeploymentStreamProvider = Literal[
     "reson8",
     "meta",
     "greenpt",
+    "typesafe",
     "slack",
     "orq",
 ]