+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -113,6 +113,7 @@ GetAllPromptsModelType = Literal[
     "rerank",
     "ocr",
     "moderation",
+    "classify",
     "vision",
 ]
 r"""The modality of the model"""
@@ -637,6 +638,7 @@ GetAllPromptsProvider = Literal[
     "reson8",
     "meta",
     "greenpt",
+    "typesafe",
     "slack",
     "orq",
 ]