+++ packages/orq-rc/RELEASES.md
@@ -11049,3 +11049,13 @@ Based on:
 - [python v4.16.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.3 - packages/orq-rc
+
+## 2026-09-18 12:15:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.4] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.4 - packages/orq-rc